Major Overhaul to Decarboxylation Progress Tracker
New Calculation to predict the rate constant for a given temperature: The formula used in the `decarboxylationTime` function calculates the slope of the line connecting two known data points (temperature and rate constant) on a plot of ln(k) versus 1/T, which is known as an Arrhenius plot.

This slope is then used to calculate the rate constant at the given temperature using the Arrhenius equation.

 The Arrhenius plot is a common tool used in chemical kinetics to analyze the temperature dependence of reaction rates. By plotting the natural logarithm of the rate constant k versus the reciprocal of the temperature 1/T (in Kelvin), the data points should lie on a straight line. The slope of this line is related to the activation energy Ea through the relationship:

    slope = -Ea / (R)

   The pre-exponential factor A can be determined from the intercept of the line with the y-axis, which corresponds to ln(A).

   Using this approach, the Arrhenius equation can be used to predict the rate constant k for a given temperature, based on known data points.